mousecam/requirements.txt
numpy
matplotlib
scipy
opencv
pyqtgraph
tqdm
click
imageio
pyqt

